Tracking of plug-in states enhanced (removed but running plug-in, name changes) + house keeping
Now all plug-in states are tracked, which was ever visible during the session.
If a plug-in is removed, but still running according to last tracking, it is still accessible to be able to stop it. (Even if it is no more listed in project setup.)
Plug-in list and changes are updated every time the project setup changes + A background check is running (currently per 1 sec), in which removal of a plug-in or plug-in name change can be detected.
During a bunch of house-keeping is done. Plug-ins got an info structure. And connecting the menu items to the plug-in info and other actions are more clear.